+++ react/admin/OrganizationSettings/index.tsx
@@ -6,19 +6,13 @@ import { useQuery, useMutation } from 'react-apollo'
 import { PageBlock, Table, IconCheck, Button, Checkbox } from 'vtex.styleguide'
 import { Spinner, useToast } from '@vtex/admin-ui'
 
-import {
-  organizationSettingsMessages as messages,
-  organizationMessages,
-} from '../utils/messages'
+import { organizationSettingsMessages as messages } from '../utils/messages'
 import { organizationBulkAction } from '../utils/organizationBulkAction'
-import GET_SALES_CHANNELS from '../../graphql/getSalesChannels.graphql'
-import SELECTED_SALES_CHANNELS from '../../graphql/getSelectedChannels.graphql'
-import UPDATE_SALES_CHANNELS from '../../graphql/updateSalesChannels.graphql'
-import UPDATE_B2B_SETTINGS from '../../graphql/updateB2BSettings.graphql'
-import GET_B2B_SETTINGS from '../../graphql/getB2BSettings.graphql'
-import GET_PAYMENT_TERMS from '../../graphql/getPaymentTerms.graphql'
-import type { PaymentTerm } from '../OrganizationDetails/OrganizationDetailsPayTerms'
-import GET_PRICE_TABLES from '../../graphql/getPriceTables.graphql'
+import GET_SALES_CHANNELS from '../graphql/getSalesChannels.graphql'
+import SELECTED_SALES_CHANNELS from '../graphql/getSelectedChannels.graphql'
+import UPDATE_SALES_CHANNELS from '../graphql/updateSalesChannels.graphql'
+import UPDATE_B2B_SETTINGS from '../graphql/updateB2BSettings.graphql'
+import GET_B2B_SETTINGS from '../graphql/getB2BSettings.graphql'
 import { TopbarCustom } from './TopbarCustom'
 
 interface SalesChannel {
@@ -51,8 +45,6 @@ const OrganizationSettings: FunctionComponent = () => {
     autoApprove: false,
     businessReadOnly: false,
     stateReadOnly: false,
-    defaultPaymentTerms: [] as any,
-    defaultPriceTables: [] as any,
     uiSettings: {
       clearCart: false,
       showModal: false,
@@ -75,12 +67,6 @@ const OrganizationSettings: FunctionComponent = () => {
     ssr: false,
   })
 
-  const { data: paymentTermsData } = useQuery<{
-    getPaymentTerms: PaymentTerm[]
-  }>(GET_PAYMENT_TERMS, { ssr: false })
-
-  const { data: priceTablesData } = useQuery(GET_PRICE_TABLES, { ssr: false })
-
   const {
     data: dataSettings,
     loading: loadingSettings,
@@ -93,35 +79,35 @@ const OrganizationSettings: FunctionComponent = () => {
   const [updateB2BSettings] = useMutation(UPDATE_B2B_SETTINGS)
 
   useEffect(() => {
-    if (data) {
-      const options: any[] = []
+    if (!data) return
 
-      data.salesChannels.forEach((item: { id: string; name: string }) => {
-        options.push({
+    const options: SalesChannel[] = []
+
+    data.salesChannels.forEach((item: { id: string; name: string }) => {
+      options.push({
+        channelId: item.id,
+        name: item.name,
+        tableName: `${item.name} (${item.id})`,
+      })
+    })
+    setSalesChannels(options)
+  }, [data])
+
+  useEffect(() => {
+    if (!selectedData) return
+
+    const selectedOptions: any[] = []
+
+    selectedData.getSalesChannels.forEach(
+      (item: { id: string; name: string }) => {
+        selectedOptions.push({
           channelId: item.id,
           name: item.name,
           tableName: `${item.name} (${item.id})`,
         })
-      })
-      setSalesChannels(options)
-    }
-  }, [data])
-
-  useEffect(() => {
-    if (selectedData) {
-      const selectedOptions: any[] = []
-
-      selectedData.getSalesChannels.forEach(
-        (item: { id: string; name: string }) => {
-          selectedOptions.push({
-            channelId: item.id,
-            name: item.name,
-            tableName: `${item.name} (${item.id})`,
-          })
-        }
-      )
-      setSelectedChannel(selectedOptions)
-    }
+      }
+    )
+    setSelectedChannel(selectedOptions)
   }, [selectedData])
 
   useEffect(() => {
@@ -135,8 +121,6 @@ const OrganizationSettings: FunctionComponent = () => {
       autoApprove: getB2BSettings?.autoApprove,
       businessReadOnly: getB2BSettings?.businessReadOnly,
       stateReadOnly: getB2BSettings?.stateReadOnly,
-      defaultPaymentTerms: getB2BSettings?.defaultPaymentTerms ?? [],
-      defaultPriceTables: getB2BSettings?.defaultPriceTables ?? [],
       uiSettings: {
         clearCart: getB2BSettings?.uiSettings?.clearCart,
         showModal: getB2BSettings?.uiSettings?.showModal,
@@ -410,74 +394,6 @@ const OrganizationSettings: FunctionComponent = () => {
               }}
               label={formatMessage(messages.fullImpersonation)}
             />
-          </div>
-          <div className="flex br3 pa6 b--muted-4 ba">
-            <div className="mb4 w-50">
-              <h2 className="mb4">
-                {formatMessage(organizationMessages.paymentTerms)}
-              </h2>
-              {paymentTermsData?.getPaymentTerms
-                .sort((a: PaymentTerm, b: PaymentTerm) => {
-                  return a.name > b.name ? 1 : -1
-                })
-                .map((payment: PaymentTerm) => (
-                  <div className="mb4">
-                    <Checkbox
-                      name={payment.name}
-                      id={payment.name}
-                      label={payment.name}
-                      checked={settings?.defaultPaymentTerms?.some(
-                        (item: PaymentTerm) => item.id === payment.id
-                      )}
-                      onChange={() => {
-                        const defaultPaymentTerms = settings.defaultPaymentTerms?.some(
-                          (item: PaymentTerm) => item.id === payment.id
-                        )
-                          ? settings.defaultPaymentTerms?.filter(
-                              (item: PaymentTerm) => item.id !== payment.id
-                            )
-                          : [...settings.defaultPaymentTerms, payment]
-
-                        setSettings({
-                          ...settings,
-                          defaultPaymentTerms,
-                        })
-                      }}
-                    />
-                  </div>
-                ))}
-            </div>
-            <div className="mb4 w-50">
-              <h2 className="mb4">
-                {formatMessage(organizationMessages.priceTables)}
-              </h2>
-              {priceTablesData?.priceTables.map((priceTable: string) => (
-                <div className="mb4">
-                  <Checkbox
-                    name={priceTable}
-                    id={priceTable}
-                    label={priceTable}
-                    checked={settings?.defaultPriceTables?.some(
-                      (item: string) => item === priceTable
-                    )}
-                    onChange={() => {
-                      const defaultPriceTables = settings.defaultPriceTables?.some(
-                        (item: string) => item === priceTable
-                      )
-                        ? settings.defaultPriceTables?.filter(
-                            (item: string) => item !== priceTable
-                          )
-                        : [...settings.defaultPriceTables, priceTable]
-
-                      setSettings({
-                        ...settings,
-                        defaultPriceTables,
-                      })
-                    }}
-                  />
-                </div>
-              ))}
-            </div>
           </div>
         </div>
       </div>