+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Remove payment terms and price table configuration
+
 ## [1.39.5] - 2024-12-12
 
 ## [1.39.4] - 2024-12-12